Table Action for the same table

+++ src/components/agenda/typecat.jsx
@@ -1,10 +1,12 @@
 import React, {Component} from "react"
 import {withRouter} from "react-router-dom";
 import {ChromePicker} from "react-color";
-import {CategoriesAgendaApi, TypesAgendaApi} from "../../helpers/request";
+import {AgendaApi, CategoriesAgendaApi, TypesAgendaApi} from "../../helpers/request";
 import EventContent from "../events/shared/content";
 import Loading from "../loaders/loading";
 import EvenTable from "../events/shared/table";
+import TableAction from "../events/shared/tableAction";
+import {handleRequestError, sweetAlert} from "../../helpers/utils";
 
 class AgendaTypeCat extends Component {
     constructor(props) {
@@ -24,7 +26,6 @@ class AgendaTypeCat extends Component {
     }
 
     async componentDidMount() {
-        console.log(this.props);
         this.eventID = this.props.event._id;
         const subject = this.props.match.url.split("/").slice(-1)[0];
         this.apiURL = subject === "categorias" ? CategoriesAgendaApi : TypesAgendaApi;
@@ -47,10 +48,10 @@ class AgendaTypeCat extends Component {
 
     newItem = () => {
         if(!this.state.list.find(({value})=>value === "new"))
-            this.setState(state => ({list: state.list.concat({label: '', value: 'new'}), id: 'new'}));
+            this.setState(state => ({list: state.list.concat({name: '', _id: 'new'}), id: 'new'}));
     };
 
-    removeNewItem = () => this.setState(state => ({list:state.list.filter(({value}) => value !== "new"),id:"",name:""}));
+    removeNewItem = () => this.setState(state => ({list:state.list.filter(({_id}) => _id !== "new"),id:"",name:""}));
 
     saveItem = async() => {
         try{
@@ -59,10 +60,9 @@ class AgendaTypeCat extends Component {
                 await this.apiURL.editOne(info, this.state.id, this.eventID);
                 this.setState(state => {
                     const data = state.list.map(object => {
-                        if (object.value === state.id) {
-                            object.item.name = state.name;
-                            object.label = state.name;
-                            object.item.color = state.color;
+                        if (object._id === state.id) {
+                            object.name = state.name;
+                            object.color = state.color;
                             return object;
                         } else return object;
                     });
@@ -72,10 +72,8 @@ class AgendaTypeCat extends Component {
                 const newRole = await this.apiURL.create(this.eventID, info);
                 this.setState(state => {
                     const types = state.list.map(item => {
-                        if (item.value === state.id) {
-                            item.label = newRole.name;
-                            item.item = newRole;
-                            item.value = newRole._id;
+                        if (item._id === state.id) {
+                            item = Object.assign(item, newRole);
                             return item;
                         } else return item;
                     });
@@ -91,15 +89,19 @@ class AgendaTypeCat extends Component {
         this.setState({id:"",name:"",color:""});
     };
 
-    removeItem = async(deleteID) => {
-        this.setState({isLoading:'Cargando....'});
-        try {
-            await this.apiURL.deleteOne(deleteID);
-            this.setState(state => ({list:state.list.filter(({value}) => value !== deleteID),id:"",name:""}))
-        }
-        catch (error) {
-            console.log(error);
-        }
+    removeItem = deleteID => {
+        sweetAlert.twoButton(`Está seguro de borrar este elemento`, "warning", true, "Borrar", async (result)=>{
+            try{
+                if(result.value){
+                    sweetAlert.showLoading("Espera (:", "Borrando...");
+                    await this.apiURL.deleteOne(deleteID);
+                    this.setState(state => ({list:state.list.filter(({_id}) => _id !== deleteID),id:"",name:""}));
+                    sweetAlert.hideLoading();
+                }
+            }catch (e) {
+                sweetAlert.showError(handleRequestError(e))
+            }
+        });
     };
 
     goBack = () => this.props.history.goBack();
@@ -111,18 +113,18 @@ class AgendaTypeCat extends Component {
                 {loading ? <Loading/> :
                     <EvenTable head={headers}>
                         {list.map(object => {
-                            return <tr key={object.value}>
+                            return <tr key={object._id}>
                                 <td>
                                     {
-                                        id === object.value ?
+                                        id === object._id ?
                                             <input type="text" value={name} autoFocus onChange={this.onChange}/>:
-                                            <p>{object.label}</p>
+                                            <p>{object.name}</p>
                                     }
                                 </td>
                                 {
                                     subject === "categorias" && <td>
                                         {
-                                            id === object.value ?
+                                            id === object._id ?
                                                 <div>
                                                     <div style={ {
                                                         padding: '5px',
@@ -143,37 +145,12 @@ class AgendaTypeCat extends Component {
                                                         <ChromePicker color={color} onChange={this.handleChangeComplete}/>
                                                     </div>}
                                                 </div>:
-                                                <div style={{padding: '5px', width: '36px', height: '14px', background: object.item.color, borderRadius: '1px', boxShadow: '0 0 0 1px rgba(0,0,0,.1)'}}/>
+                                                <div style={{padding: '5px', width: '36px', height: '14px', background: object.color, borderRadius: '1px', boxShadow: '0 0 0 1px rgba(0,0,0,.1)'}}/>
                                         }
                                     </td>
                                 }
-                                <td>
-                                    {
-                                        id === object.value ?
-                                            <button data-tooltip="Guardar">
-                                            <span className="icon has-text-grey"
-                                                  onClick={(e)=>{this.saveItem(object)}}><i className="fas fa-save"/></span>
-                                            </button>:
-                                            <button data-tooltip="Editar">
-                                            <span className="icon has-text-grey"
-                                                  onClick={(e)=>this.editItem(object.item)}><i className="fas fa-edit"/></span>
-                                            </button>
-                                    }
-                                    {
-                                        object.value === 'new' ?
-                                            <button data-tooltip="Eliminar"><span className='icon has-text-grey'
-                                                          onClick={this.removeNewItem}><i className='fas fa-times'/></span></button>:
-                                            <button data-tooltip="Eliminar"><span className='icon has-text-grey'
-                                                          onClick={(e)=>{this.removeItem(object.value)}}><i className='far fa-trash-alt'/></span></button>
-                                    }
-                                    {
-                                        id === object.value &&
-                                        <button data-tooltip="Cancelar">
-                                            <span className="icon has-text-grey"
-                                                  onClick={this.discardChanges}><i className="fas fa-times"/></span>
-                                        </button>
-                                    }
-                                </td>
+                                <TableAction id={id} object={object} saveItem={this.saveItem} editItem={this.editItem}
+                                    removeNew={this.removeNewItem} removeItem={this.removeItem} discardChanges={this.discardChanges}/>
                             </tr>
                         })}
                     </EvenTable>
